+++ R/data-sets.R
@@ -39,7 +39,8 @@
 #' phx
 #'
 #' A synthetic dataset for three subjects' electronic health record related to
-#' a full encounter.
+#' a full encounter.  This is a "long" format dataset used for examples of the
+#' operationalized workflow.
 #'
 #' @format a data.frame with four columns:
 #'
@@ -54,6 +55,7 @@
 #' column is numeric, when the following table says the value type is an
 #' integer, that denotes the practical use of the variable, not the storage
 #' mode.
+#'
 #' \tabular{llll}{
 #'  \strong{Name}                          \tab \strong{Common Abbreviation(s)} \tab \strong{Value type} \tab \strong{Units}    \cr
 #'  absolute_lymphocyte_count              \tab ALC                             \tab numeric             \tab 1e3 cells per mm³ \cr
@@ -61,8 +63,8 @@
 #'  age_months                             \tab                                 \tab numeric             \tab months            \cr
 #'  alanine_aminotransferase               \tab ALT                             \tab numeric             \tab IU/L              \cr
 #'  creatinine                             \tab                                 \tab numeric             \tab mg/dL             \cr
-#'  diastolic_blood_pressure_arterial_line \tab MAP_ART                         \tab numeric             \tab mmHg              \cr
-#'  diastolic_blood_pressure_cuff          \tab MAP_CUFF                        \tab numeric             \tab mmHg              \cr
+#'  diastolic_blood_pressure_arterial_line \tab DBP_ART                         \tab numeric             \tab mmHg              \cr
+#'  diastolic_blood_pressure_cuff          \tab DBP_CUFF                        \tab numeric             \tab mmHg              \cr
 #'  fibrin_degradation_fragment            \tab D-Dimer; ddimer                 \tab numeric             \tab mg/L FEU          \cr
 #'  fibrinogen                             \tab                                 \tab numeric             \tab mg/dL             \cr
 #'  fraction_of_inspired_oxygen            \tab FiO2                            \tab numeric             \tab                   \cr
@@ -72,7 +74,7 @@
 #'  glasgow_coma_scale_verbal              \tab GCS_VERBAL                      \tab integer             \tab                   \cr
 #'  glucose                                \tab                                 \tab numeric             \tab mg/dL             \cr
 #'  infectious_test                        \tab                                 \tab integer (indicator) \tab                   \cr
-#'  internation_normalized_ratio           \tab INR                             \tab numeric             \tab                   \cr
+#'  international_normalized_ratio         \tab INR                             \tab numeric             \tab                   \cr
 #'  lactate                                \tab                                 \tab numeric             \tab mmol/L            \cr
 #'  left_pupil_fixed                       \tab                                 \tab integer (indicator) \tab                   \cr
 #'  mean_airway_pressure_ventilator        \tab MAP_VENT                        \tab numeric             \tab cmH2O             \cr
@@ -80,15 +82,15 @@
 #'  mean_arterial_pressure_cuff            \tab MAP_CUFF                        \tab numeric             \tab mmHg              \cr
 #'  oxygen_flow                            \tab                                 \tab numeric             \tab L/min             \cr
 #'  partial_pressure_of_arterial_oxygen    \tab PaO2                            \tab numeric             \tab mmHg              \cr
-#'  peripheral_capillary_oxygen_saturation \tab SpO2                            \tab numeric             \tab mmHg              \cr
+#'  peripheral_capillary_oxygen_saturation \tab SpO2                            \tab numeric             \tab percent           \cr
 #'  platelets                              \tab PLTS                            \tab numeric             \tab 1000/uL (1000 per micro-liter) \cr
 #'  positive_end_expiratory_pressure       \tab PEEP; PEEP_VENT                 \tab numeric             \tab cm H2O            \cr
 #'  right_pupil_fixed                      \tab                                 \tab integer (indicator) \tab                   \cr
 #'  systemic_antimicrobial_medication      \tab                                 \tab integer (indicator) \tab                   \cr
 #'  systemic_vasoactive_epinephrine        \tab                                 \tab integer (indicator) \tab                   \cr
 #'  systemic_vasoactive_milrinone          \tab                                 \tab integer (indicator) \tab                   \cr
-#'  systolic_blood_pressure_arterial_line  \tab                                 \tab mmHg                \tab                   \cr
-#'  systolic_blood_pressure_cuff           \tab                                 \tab mmHg                \tab                   \cr
+#'  systolic_blood_pressure_arterial_line  \tab                                 \tab numeric             \tab mmHg              \cr
+#'  systolic_blood_pressure_cuff           \tab                                 \tab numeric             \tab mmHg              \cr
 #'  total_bilirubin                        \tab BILIRUBIN_TOT                   \tab numeric             \tab mg/dL             \cr
 #' }
 "phx"